requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Base de Datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Procesamiento de Datos / EDA
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib>=3.8.0

# NLP / Análisis de Sentimiento
transformers==4.35.2
torch>=2.2.0
sentencepiece==0.1.99

# Utilidades
pydantic==2.5.2
pydantic-settings==2.1.0
